Returning HTTP Error 404 (Not Found) when appropriated.

+++ src/CatalogAPI/Catalog.Service/Books/Handler/BookDeleteRequestHandler.cs
@@ -1,12 +1,13 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
 using Catalog.Domain.Entity;
+using Catalog.Service.Books.Request;
 using Liquid.Repository;
 using MediatR;
 
 namespace Catalog.Service.Books.Handler
 {
-    public sealed class BookDeleteRequestHandler : IRequestHandler<Request.BookDeleteRequest>
+    public sealed class BookDeleteRequestHandler : IRequestHandler<Request.BookDeleteRequest, Book>
     {
         private readonly ILiquidRepository<Book, string> _booksRepository;
 
@@ -15,11 +16,15 @@ namespace Catalog.Service.Books.Handler
             _booksRepository = booksRepository;
         }
 
-        public async Task<Unit> Handle(Request.BookDeleteRequest request, CancellationToken cancellationToken)
+        public async Task<Book> Handle(Request.BookDeleteRequest request, CancellationToken cancellationToken)
         {
-            await _booksRepository.RemoveByIdAsync(request.Id);
+            var book = await _booksRepository.FindByIdAsync(request.Id);
+            if (book != null)
+            {
+                await _booksRepository.RemoveByIdAsync(request.Id);
+            }
 
-            return new Unit();
+            return book;
         }
     }
 }
